Set default _unicastPort to 0, allowing it to vary
Use the "-u port" option to fix it to a given port.

This however, brings up another issue, in that the server
may still be sending to that fixed port until it notices the
TCP port is gone via the keepalive trick.  So it will be sending
data instead of the initialization packets handshake, which are
screened by checking the magic number. This situation
could be reduced somewhat by using a varying port for the
initialization.  Since the initialization packets are going to a known
port on the server, they can be allowed in a firewall.